split out char validation
