+++ src/lib.rs
@@ -3,6 +3,7 @@
 #![forbid(unsafe_code)]
 #![doc = include_str!("../README.md")]
 
+mod char_validator;
 mod emitter;
 mod entities;
 mod error;
